requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
Django==4.0.7
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.4.0
importlib-metadata==4.12.0
Markdown==3.4.1
Pillow==9.2.0
PyJWT==2.4.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
zipp==3.8.1
